+++ src/components/Pill/index.js
@@ -1,3 +1,0 @@
-export { PillButton } from './PillButton';
-export { PillDropdown } from './PillDropdown';
-export { Pill } from './Pill';