+++ stories/molecules/Pill.tsx
@@ -12,6 +12,10 @@ export default {
     title: 'Molecules/Pill',
     component: Pill,
     subcomponents: { PillButton, PillDropdown },
+    argTypes: {
+        ref: { control: false },
+        dropdownRef: { control: false },
+    },
 };
 
 export const _Pill = (args) => {
@@ -31,7 +35,7 @@ export const _PillButton = (args) => (
     <div style={{ display: 'flex' }}>
         <PillButton {...args} />
         &nbsp;&nbsp;
-        <PillButton name="Pill 2" content="" />
+        <PillButton {...args} name="Pill 2" content="" />
     </div>
 );
 _PillButton.args = {